+++ packages/charts/src/components/CarbonChart/CarbonChartBarGrouped.stories.tsx
@@ -28,6 +28,130 @@ const Template: Story<CarbonChartProps> = args => (
 );
 
 //Grouped//
+export const PreSelectedgroupsGroupedBar = Template.bind({});
+PreSelectedgroupsGroupedBar.args = {
+  isInverse: false,
+  type: CarbonChartType.barGrouped,
+  dataSet: [
+    {
+      group: 'Dataset 1',
+      key: 'Qty',
+      value: 65000,
+    },
+    {
+      group: 'Dataset 1',
+      key: 'More',
+      value: -29123,
+    },
+    {
+      group: 'Dataset 1',
+      key: 'Sold',
+      value: -35213,
+    },
+    {
+      group: 'Dataset 1',
+      key: 'Restocking',
+      value: 51213,
+    },
+    {
+      group: 'Dataset 1',
+      key: 'Misc',
+      value: 16932,
+    },
+    {
+      group: 'Dataset 2',
+      key: 'Qty',
+      value: 32432,
+    },
+    {
+      group: 'Dataset 2',
+      key: 'More',
+      value: -21312,
+    },
+    {
+      group: 'Dataset 2',
+      key: 'Sold',
+      value: -56456,
+    },
+    {
+      group: 'Dataset 2',
+      key: 'Restocking',
+      value: -21312,
+    },
+    {
+      group: 'Dataset 2',
+      key: 'Misc',
+      value: 34234,
+    },
+    {
+      group: 'Dataset 3',
+      key: 'Qty',
+      value: -12312,
+    },
+    {
+      group: 'Dataset 3',
+      key: 'More',
+      value: 23232,
+    },
+    {
+      group: 'Dataset 3',
+      key: 'Sold',
+      value: 34232,
+    },
+    {
+      group: 'Dataset 3',
+      key: 'Restocking',
+      value: -12312,
+    },
+    {
+      group: 'Dataset 3',
+      key: 'Misc',
+      value: -34234,
+    },
+    {
+      group: 'Dataset 4',
+      key: 'Qty',
+      value: -32423,
+    },
+    {
+      group: 'Dataset 4',
+      key: 'More',
+      value: 21313,
+    },
+    {
+      group: 'Dataset 4',
+      key: 'Sold',
+      value: 64353,
+    },
+    {
+      group: 'Dataset 4',
+      key: 'Restocking',
+      value: 24134,
+    },
+    {
+      group: 'Dataset 4',
+      key: 'Misc',
+      value: 24134,
+    },
+  ],
+  options: {
+    title: 'Pre-selected groups (grouped bar)',
+    data: {
+      selectedGroups: ['Dataset 1', 'Dataset 3'],
+    },
+    axes: {
+      left: {
+        mapsTo: 'value',
+      },
+      bottom: {
+        scaleType: 'labels',
+        mapsTo: 'key',
+      },
+    },
+    height: '400px',
+  },
+};
+
 export const VerticalGroupedBarDiscrete = Template.bind({});
 VerticalGroupedBarDiscrete.args = {
   isInverse: false,