+++ packages/react-magma-dom/src/components/Select/utils.ts
@@ -1,0 +1,5 @@
+import { instanceOfItemWithOptionalDisabled } from ".";
+
+export function isItemDisabled(item) {
+  return instanceOfItemWithOptionalDisabled(item) && item.disabled;
+}